add util for pulling time data from form
